requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
ecdsa==0.18.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
icecream==2.1.3
idna==3.4
mysql-connector-python==8.2.0
passlib==1.7.4
protobuf==4.21.12
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.1
pydantic-settings==2.1.0
pydantic_core==2.14.3
Pygments==2.17.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
